requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-RESTful==0.3.9
flask-swagger-ui==3.36.0
Flask-Migrate==3.1.0
PyPDF2
python-docx
openpyxl
pandas
matplotlib
plotly
